requirements.txt
chalice
robotframework==3.2.2
robotframework-seleniumlibrary==4.5.0
boto3==1.14.63